requirements.txt
asgiref==3.2.10
Django==3.1.1
django-debug-toolbar==2.2
gunicorn==20.0.4
importlib-metadata==1.7.0
Markdown==3.2.2
Pillow==7.2.0
psycopg2-binary==2.8.6
pytz==2020.1
sqlparse==0.3.1
zipp==3.1.0
environ==1.0
